refactor: deleting unnecessary files
