+++ packages/stylelint-plugin-slds/tests/rules/no-important/no-important.spec.ts
@@ -2,7 +2,8 @@
 import stylelint, { LinterResult, LinterOptions } from 'stylelint';
 const { lint }: typeof stylelint = stylelint;
 
-describe('no-important', () => {
+//ToDo: remove once tests migrated to eslint completely
+xdescribe('no-important', () => {
   [
     "Avoid using '!important' unless absolutely necessary.",
     "Avoid using '!important' unless absolutely necessary.",
