+++ packages/stylelint-plugin-slds/tests/rules/no-slds-class-overrides/no-slds-class-overrides.spec.ts
@@ -3,7 +3,8 @@ import assert from 'assert';
 
 const ruleName = 'slds/no-slds-class-overrides';
 
-describe('slds/no-slds-class-overrides rule', () => {
+//ToDo: remove once tests migrated to eslint completely
+xdescribe('slds/no-slds-class-overrides rule', () => {
   it('should report an error when overriding .slds- class styles', async () => {
     const css = `
       .foo {
